# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAT API
  version: 1.0.0
  description: Contrato inicial da API do SIGAT, alinhado aos fluxos consumidos pelo frontend.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API versionada atrás do Nginx
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Cria uma sessão autenticada
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Contexto da sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContext'
        '400':
          $ref: '#/components/responses/InvalidJson'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalida a sessão atual
      operationId: logout
      security: [{ sessionCookie: [] }]
      responses:
        '204': { description: Sessão invalidada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/password-change:
    post:
      tags: [Auth]
      summary: Troca a senha do próprio usuário quando há troca obrigatória pendente
      operationId: changeOwnPassword
      description: Usado após login com `mustChangePassword=true`. Aceita somente a nova senha, grava apenas hash bcrypt e retorna o contexto de sessão já liberado.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordChangeRequest' }
      responses:
        '200':
          description: Senha alterada e sessão liberada
          content: { application/json: { schema: { $ref: '#/components/schemas/SessionContext' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
  /session:
    get:
      tags: [Auth]
      summary: Consulta a sessão atual
      operationId: getSession
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Contexto autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContext'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /audit-events:
    get:
      tags: [Audit]
      summary: Lista eventos de auditoria
      operationId: listAuditEvents
      description: Exclusivo para administrador DITEL. Retorna eventos do mais recente para o mais antigo, sem credenciais, tokens ou identificadores de sessão.
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: module, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: userId, in: query, schema: { type: string } }
        - { name: unitId, in: query, schema: { type: string } }
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: string } }
        - { name: result, in: query, schema: { type: string, enum: [success, failure] } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Página de eventos de auditoria
          content: { application/json: { schema: { $ref: '#/components/schemas/AuditEventListResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /calls/{callId}:
    get:
      tags: [Calls]
      summary: Consulta detalhes públicos de um chamado
      operationId: getCallDetails
      description: Retorna dados operacionais públicos do chamado e histórico derivado de dados persistidos e auditoria. Administradores DITEL podem consultar chamados estaduais; usuários de unidade recebem apenas chamados da própria unidade.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: callId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detalhes públicos do chamado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CallDetails' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /calls/{callId}/triage:
    patch:
      tags: [Calls]
      summary: Persiste a triagem operacional de um chamado
      operationId: triageCall
      description: Exclusivo para administrador DITEL. Atualiza somente seção responsável, prioridade e situação permitidas, usando updatedAt como token de concorrência; versões antigas retornam 409 e a auditoria obrigatória pode reverter a mutação.
      security: [{ sessionCookie: [] }]
      parameters:
        - name: callId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TriageRequest' }
      responses:
        '200':
          description: Chamado atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CallQueueItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
...
  /calls:
    get:
      tags: [Calls]
      summary: Lista chamados dentro do escopo autorizado
      operationId: listCalls
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Fila de chamados
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CallQueueItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Calls]
      summary: Abre um chamado
      operationId: createCall
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CallRequest'
      responses:
        '201':
          description: Chamado criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedCall'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/AuditUnavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sigat_session
  parameters:
...
    Page: { name: page, in: query, required: true, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    BadRequest:
      description: Dados inválidos
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    InvalidJson:
      description: Corpo JSON malformado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            malformedJson:
              value:
                code: INVALID_JSON
                message: Corpo JSON inválido.
    Unauthorized:
      description: Sessão ausente ou inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Usuário autenticado sem uma unidade válida para acessar o dashboard da unidade
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            invalidUnitScope:
              value:
                code: FORBIDDEN
                message: Escopo de unidade obrigatório.
    NotFound:
      description: Recurso não encontrado no escopo autorizado
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflito de unicidade ou concorrência
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    AuditUnavailable:
      description: A mutação foi revertida porque a auditoria obrigatória não pôde ser registrada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            auditUnavailable:
              value:
                code: AUDIT_UNAVAILABLE
                message: Não foi possível registrar auditoria.
  schemas:
...
    AuditActor:
      type: object
      required: [id]
      properties:
        id: { type: string }
        name: { type: string }
        registration: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
    AuditEntity:
      type: object
      required: [type, id]
      properties:
        type: { type: string }
        id: { type: string }
        label: { type: string }
    AuditSnapshot:
      oneOf:
        - type: object
          additionalProperties: true
        - type: 'null'
    AuditEvent:
      type: object
      required: [id, action, module, userId, actor, entity, unit, result, reason, before, after, retentionExpiresAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        action: { type: string }
        module: { type: string }
        userId:
          oneOf:
            - { type: string }
            - { type: 'null' }
        actor:
          oneOf:
            - { $ref: '#/components/schemas/AuditActor' }
            - { type: 'null' }
        entity:
          oneOf:
            - { $ref: '#/components/schemas/AuditEntity' }
            - { type: 'null' }
        unit:
          oneOf:
            - { $ref: '#/components/schemas/UnitContext' }
            - { type: 'null' }
        result: { type: string, enum: [success, failure] }
        reason:
          oneOf:
            - { type: string }
            - { type: 'null' }
        before: { $ref: '#/components/schemas/AuditSnapshot' }
        after: { $ref: '#/components/schemas/AuditSnapshot' }
        retentionExpiresAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AuditEventListResponse:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: EQUIPMENT_NOT_FOUND }
        message: { type: string }
        details: {}
...
    LoginRequest:
      type: object
      required: [registration, password]
      properties:
        registration: { type: string }
        password: { type: string, format: password }
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [newPassword]
      properties:
        newPassword: { type: string, minLength: 8, format: password, writeOnly: true }
    UserRole:
      type: string
      enum: [ditel_admin, unit_user]
...
    UnitContext:
      type: object
      required: [id, name, acronym]
      properties:
        id: { type: string }
        name: { type: string }
        acronym: { type: string }
    SessionContext:
      type: object
      required: [userId, name, registration, role, unit]
      properties:
        userId: { type: string }
        name: { type: string }
        registration: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        mustChangePassword:
          type: boolean
          default: false
          description: Quando verdadeiro, o frontend deve direcionar o usuário para troca obrigatória de senha antes de continuar.
        unit:
          oneOf:
            - { $ref: '#/components/schemas/UnitContext' }
            - { type: 'null' }
...
    HistoryEntry:
      type: object
      required: [id, description, occurredAt]
      properties:
        id: { type: string }
        description: { type: string }
        occurredAt: { type: string }
...
    AttachmentMetadata:
      type: object
      required: [id, name, type, size, uploadedAt, status, downloadUrl]
      additionalProperties: false
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        size: { type: integer, minimum: 0 }
        uploadedAt: { type: string, format: date-time }
        status: { type: string, enum: [active] }
        downloadUrl: { type: string, format: uri-reference }
...
    CallQueueItem:
      type: object
      required: [id, subject, unitId, unitName, priority, status, updatedAt]
      properties:
        id: { type: string }
        subject: { type: string }
        unitId: { type: string }
        unitName: { type: string }
        priority: { type: string, enum: [Baixa, Média, Alta, Crítica], description: Prioridade exibida retornada pela API. }
        status: { type: string, enum: [Aberto, Em análise, Em atendimento, Aguardando informação, Resolvido, Encerrado] }
        section: { type: string, enum: [Suporte, Telecom] }
        updatedAt: { type: string, format: date-time }
    CallDetailEquipment:
      oneOf:
        - type: object
          required: [id, patrimony, type, model, brand]
          properties:
            id: { type: string }
            patrimony: { type: string }
            type: { type: string }
            model: { type: string }
            brand: { type: string }
        - { type: 'null' }
    CallDetails:
      type: object
      required: [id, protocol, problem, subject, description, unit, requestedBy, priority, status, section, equipment, openedAt, updatedAt, attachments, history]
      properties:
        id: { type: string }
        protocol: { type: string }
        problem: { type: string, enum: [software, hardware, printer, network, radio] }
        subject: { type: string }
        description: { type: string }
        unit: { $ref: '#/components/schemas/UnitContext' }
        requestedBy: { type: string }
        priority: { type: string, enum: [Baixa, Média, Alta, Crítica] }
        status: { type: string, enum: [Aberto, Em análise, Em atendimento, Aguardando informação, Resolvido, Encerrado] }
        section: { type: string, enum: [Suporte, Telecom] }
        equipment: { $ref: '#/components/schemas/CallDetailEquipment' }
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        attachments:
          description: Metadados públicos de anexos persistidos; não inclui storageKey, caminho local ou segredos.
          type: array
          items: { $ref: '#/components/schemas/AttachmentMetadata' }
        history: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } }
    TriageRequest:
      type: object
      required: [status, priority, section, updatedAt]
      additionalProperties: false
      properties:
        status: { type: string, enum: [Em análise, Em atendimento, Aguardando informação, Resolvido, Encerrado] }
        priority: { type: string, enum: [low, medium, high, critical] }
        section: { type: string, enum: [Suporte, Telecom] }
        updatedAt: { type: string, format: date-time }
    CallRequest:
      type: object
      required: [problem, priority, subject, description]
      properties:
        problem: { type: string, enum: [software, hardware, printer, network, radio] }
        priority: { type: string, enum: [low, medium, high, critical], description: Prioridade técnica enviada pelo frontend. }
        subject: { type: string, maxLength: 120 }
        description: { type: string }
        equipmentId: { type: string }
        attachments:
          description: Até 5 arquivos, validados por MIME e extensão. Aceita PDF/DOC/DOCX e imagens JPG/PNG/WEBP/GIF, com limite de 10 MiB por arquivo.
          type: array
          maxItems: 5
          items:
            type: string
            format: binary
    CreatedCall:
      type: object
      required: [id, protocol]
      properties:
        id: { type: string }
        protocol: { type: string }